+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/google-vertex
 
+## 3.0.65
+
+### Patch Changes
+
+- 9a66892: Add Google Vertex RAG Engine grounding provider tool
+- Updated dependencies [9a66892]
+  - @ai-sdk/google@2.0.34
+
+## 3.0.64
+
+### Patch Changes
+
+- Updated dependencies [30a7d54]
+  - @ai-sdk/google@2.0.33
+
+## 3.0.63
+
+### Patch Changes
+
+- Updated dependencies [0ac13cc]
+  - @ai-sdk/google@2.0.32
+
 ## 3.0.62
 
 ### Patch Changes