+++ packages/google/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/google
 
+## 2.0.34
+
+### Patch Changes
+
+- 9a66892: Add Google Vertex RAG Engine grounding provider tool
+
+## 2.0.33
+
+### Patch Changes
+
+- 30a7d54: Improve error message when mixing function tools with provider-defined tools to clarify fallback behavior and list ignored function tools
+
+## 2.0.32
+
+### Patch Changes
+
+- 0ac13cc: feat(vertex): support 'trafficType' in provider usageMetadata
+
 ## 2.0.31
 
 ### Patch Changes